The-Food-Court-master/backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.7.0
djangorestframework==3.12.4
Faker==9.3.1
future==0.18.2
idna==3.3
mysqlclient==2.0.3
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
urllib3==1.26.7
factory-boy
